Former-commit-id: 5cc9d364c49f3656e6ca69f91915ff465a950479

+++ Sources/Internal/Scene3D/Components/SkeletonComponent.cpp
@@ -71,7 +71,7 @@ SkeletonComponent::JointConfig::JointConfig(int32 _parentIndex, int32 _targetId,
 
 uint16 SkeletonComponent::GetConfigJointsCount()
 {
-    return configJoints.size();
+    return uint16(configJoints.size());
 }
 void SkeletonComponent::SetConfigJoints(const Vector<JointConfig>& config)
 {
